+++ src/screen/signIn/api/index.js
@@ -1,15 +1,4 @@
-import axios from ".";
-
-export const addNewUser = async (newUser) => {
-  console.log(newUser);
-  try {
-    const response = await axios.post("/register-user/", newUser);
-    console.log(response);
-    return response;
-  } catch (error) {
-    return error;
-  }
-};
+import axios from "../../../api";
 
 export const login = async (user) => {
   try {
